requirements.txt
tweepy>=4.14.0
pytube>=15.0.0
replicate>=0.22.0
flask>=3.0.0
markdown>=3.5.1
python-dotenv>=1.0.0